admin app: booking request

+++ backend/src/main/java/com/example/bokningsapp/repository/HubRepository.java
@@ -4,7 +4,9 @@ import com.example.bokningsapp.model.Hub;
 import com.example.bokningsapp.model.bookings.HubBooking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HubRepository extends JpaRepository<Hub, Integer> {
+import java.util.List;
+
+public interface HubRepository extends JpaRepository<Hub, Long> {
 
 
 }
